create directory templates and move templates
