+++ js/templates/attemptsEndedModule.js
