+++ js/templates/resultSuccessModule.js
