+++ js/templates/welcomeModule.js
